Fix state field value when a country is prefilled. Extract types to its own file. Update import path.
